pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "text_to_image"
version = "0.0.1-beta"                                                    # or dynamic from git?
authors = [{ name = "Benjamin Liden", email = "lidenb@oregonstate.edu" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">3.10"
dependencies = [
    "pillow", 
    "pydantic", 
    "pydantic-settings",
    "zmq"
]

[project.urls]
Homepage = "https://github.com/bcliden/text_to_image"
Issues = "https://github.com/bcliden/text_to_image/issues"

[tool.setuptools.package-data]
"text_to_image.font" = ["*.ttf"]

[project.scripts]
"text_to_image" = "text_to_image.zero_mq.entry:main"

[project.gui-scripts]
# maybe just here
